# New design for better performance of no batching case with multiple i…