--- a/src/backends/backend/triton_model.h
+++ b/src/backends/backend/triton_model.h
@@ -75,9 +75,6 @@ class TritonModel : public InferenceBackend {
       std::unique_ptr<TritonModelInstance>&& instance, const bool passive,
       const inference::ModelRateLimiter& rate_limiter_config);
 
-  Status Initialize();
-  Status WarmUp();
-
  private:
   DISALLOW_COPY_AND_ASSIGN(TritonModel);
 
@@ -86,6 +83,9 @@ class TritonModel : public InferenceBackend {
       const std::shared_ptr<LocalizedDirectory>& localized_model_dir,
       const std::shared_ptr<TritonBackend>& backend,
       const double min_compute_capability, const bool auto_complete_config);
+
+  Status Initialize();
+  Status WarmUp();
 
   // The server object that owns this model. The model holds this as a
   // raw pointer because the lifetime of the server is guaranteed to