…nstances

So, in this re-design we have modified the interface of rate limiter quite a bit. Instead of earlier callback based signaling, we will use the Enqueue and Dequeue methods to control when each instance gets to run inference.

The current implementation is just a proposal for a simple no batching case. I have verified that it keeps the same performance as our current threading model when the RL is disabled. However, there are some weirdness that I can not yet explain. More on this at the end.

Following are some points about this design:
1. The scheduler threads will just be batcher threads and will come into the picture only when a batching is enabled.
2. The request queue is maintained by RateLimiter. This will be replaced with PriorityQueue from scheduler_utils in the future to include functionalities like timeouts and priorities.
3. BackendThreads will wait on RL to receive a payload to use for inference. In case of the high load and no rate limiting, this reduces to just a pop out of the queue.
4. Payload infrastructure is kept because later for device blocking case the same thread can be used to intialize and warmup the model instances whenever required.
5. In order to interact with the BackendThread, user must request for a payload oobject from RL, configure the object and then enqueue to the RL. The RL will then pass it to the waiting BackendThread whenever scheduling for inference.

**There are some issues I am trying to resolve:**
Once in a while in PA I see a window does not have any valid requests. This is strange because I see executions happening all the time without ay 5 sec delay. I would need to revisit the PA logic and infer_stats.cc to see why there are no valid requests recorded in the window. 
```
  Pass [1] throughput: 1633.8 infer/sec. Avg latency: 2448 usec (std 8578 usec)
  Pass [2] throughput: 816 infer/sec. Avg latency: 2450 usec (std 12551 usec)
  Pass [3] throughput: 1634.4 infer/sec. Avg latency: 2457 usec (std 7721 usec)
  Pass [4] throughput: 1629.4 infer/sec. Avg latency: 2462 usec (std 7706 usec)
  Pass [5] cb::Error: No valid requests recorded within time interval. Please use a larger time window.
  Pass [6] throughput: 1633 infer/sec. Avg latency: 2439 usec (std 7129 usec)
  Pass [7] throughput: 300.4 infer/sec. Avg latency: 2430 usec (std 9170 usec)
  Pass [8] throughput: 1630 infer/sec. Avg latency: 2453 usec (std 8582 usec)
  Pass [9] throughput: 1632 infer/sec. Avg latency: 2456 usec (std 7829 usec)
```