@deadeyegoodwin 2021-06-10
> 
> 
> I do have to implement the preserve ordering case as well.

It seems like that is completely handled by the batcher and the inference request... I wouldn't expect your changes to have any interaction